moved style.css file out of static folder
style.css doesn't seem to do much when removed. but the line of code refering to it on layout.html seem to affect it. Thus removing the "style.css" file, toring it in the "Tests" folder with a different anem "BelongsInStaticFolder_but_DoesNotLookRequired_style.css" to see what would happen